Added ReduceRule, for mapping tokens to another form
The motivation for this was to aggregate multiple tokens into a single
token. However, reduce can map a sequence of tokens into an arbitrary
sequence, so it's potentially much more powerful than just a single
aggregation.

ReduceRule can also remap to entirely different tokens, so there's no
need to keep everything working in the same token type if this isn't
needed.