Several major API changes
All rules parse method has changed to:

bool parse(Cursor<Input>&, Result<Input>&)

This is primarily done for performance purposes. The old version created
unnecessary Results for all matches and non-matches. This forced a large
amount of copying to get the result. Changing to a single result value
doubled the performance.

I also changed how makeCursor works with regard to strings of all types.
The old version converted all strings to streams and iterated over the
stream. This turned out not to be necessary, as strings can act as their
own iterators. This increases speed for our tests and benchmarks, which
contributes to the speed increase mentioned above.