Don't use shared_ptr when working with results
This slightly limits how flexible Results can be, but since we've
changed how we treat them already (to creating a minimal amount of
them), this change isn't as impactful as it initially appeared.

As an aside, this change means Sprout is faster than Qt's regular
expressions for all current tests (and using GCC).