Lots of refactors to SequenceRule
First, I removed the "Optional" stream manipulators, as the new rule
function signature is sufficient to keep that behavior external to
Sequence. Removing this data allowed me to remove the RuleEntry struct
entirely, which flattens a bunch of code.

I also changed SequenceRule to behave like the other composite rules. As
a result, I was able to use the composite.hpp's populate() instead of
rolling my own here.