Added SharedRule
Normally, rules are copied when they are composed together. This works
well for efficiency, but sometimes it's required to retain access to a
rule to fully specify it. This can occur if there is a cylical dependency
between two rules (e.g. a list that can contain rvalues, and rvalues that
can be lists).