Added two optimizations for Result
The first is a suppress flag, which causes a Result to ignore any inserted
values. This ensures rules that ignore output, like DiscardRule, don't
need to create then immediately waste creating a Result.

The second optimization is adding an insertion position or "head". Clients
can reposition the head if they want to disregard results. This allows
SequenceRule to avoid creating a temporary Result that will eventually
be added.

These two optimizations are significant in rules that use a lot of discards
and sequences, which most seem to do.