Major renaming and reorganization
The motivation for this is to make room for the new Grammar stuff. I'd
like to lift this out of src/ and integrate it fully into the rest of the
library. I wanted to keep things separate, and organized by namespace,
hence the large amount of moves. I also removed the *Rule suffix for all
rules, so now you refer to them as rule::Proxy rather than ProxyRule.